wip: start moving confirm delet popup to main
